+++ inc/footer/footer-main-menu.php
@@ -1,9 +1,9 @@
 <?php
 /**
- * Hardcoded Footer Menu
+ * Footer Main Menu
  *
- * Clean, maintainable footer menu structure without WordPress menu system overhead.
- * Direct HTML control for optimal performance and easy maintenance.
+ * Network-centric footer navigation structure.
+ * Clean HTML structure without WordPress menu system overhead.
  *
  * @package ExtraChill
  * @since 1.0.0
@@ -14,75 +14,47 @@
     <div class="footer-menu-column">
         <ul class="footer-column-menu">
             <li class="menu-item menu-item-has-children">
-                <a href="https://extrachill.com/all">The Latest</a>
+                <a href="https://extrachill.com">Network</a>
                 <ul class="sub-menu">
                     <li class="menu-item">
-                        <a href="https://extrachill.com/festival-wire">Festival Wire</a>
+                        <a href="https://extrachill.com/all">Blog</a>
                     </li>
                     <li class="menu-item">
-                        <a href="https://events.extrachill.com">Events Calendar</a>
+                        <a href="https://community.extrachill.com">Community</a>
                     </li>
                     <li class="menu-item">
-                        <a href="https://extrachill.com/category/interviews/">Interviews</a>
-                    </li>
-                    <li class="menu-item">
-                        <a href="https://extrachill.com/category/live-music-reviews/">Live Music Reviews</a>
-                    </li>
-                    <li class="menu-item">
-                        <a href="https://extrachill.com/category/music-news/">Music News</a>
-                    </li>
-                    <li class="menu-item">
-                        <a href="https://newsletter.extrachill.com">Newsletters</a>
-                    </li>
-                </ul>
-            </li>
-        </ul>
-    </div>
-
-    <div class="footer-menu-column">
-        <ul class="footer-column-menu">
-            <li class="menu-item menu-item-has-children">
-                <a href="#">The Rabbit Hole</a>
-                <ul class="sub-menu">
-                    <li class="menu-item">
-                        <a href="https://extrachill.com/category/song-meanings/">Song Meanings</a>
-                    </li>
-                    <li class="menu-item">
-                        <a href="https://extrachill.com/artist/grateful-dead">Grateful Dead</a>
-                    </li>
-                    <li class="menu-item">
-                        <a href="https://extrachill.com/category/musical-curiosities/">Musical Curiosities</a>
-                    </li>
-                    <li class="menu-item">
-                        <a href="https://extrachill.com/category/famous-guitars/">Famous Guitars</a>
-                    </li>
-                </ul>
-            </li>
-        </ul>
-    </div>
-
-    <div class="footer-menu-column">
-        <ul class="footer-column-menu">
-            <li class="menu-item menu-item-has-children">
-                <a href="https://community.extrachill.com/">Community</a>
-                <ul class="sub-menu">
-                    <li class="menu-item">
-                        <a href="https://community.extrachill.com/recent">Recent Activity</a>
+                        <a href="https://events.extrachill.com">Events</a>
                     </li>
                     <li class="menu-item">
                         <a href="https://artist.extrachill.com">Artists</a>
                     </li>
                     <li class="menu-item">
-                        <a href="https://community.extrachill.com/r/local-scenes">Local Scenes</a>
+                        <a href="https://newsletter.extrachill.com">Newsletter</a>
+                    </li>
+                </ul>
+            </li>
+        </ul>
+    </div>
+
+    <div class="footer-menu-column">
+        <ul class="footer-column-menu">
+            <li class="menu-item menu-item-has-children">
+                <a href="https://extrachill.com/all">Explore</a>
+                <ul class="sub-menu">
+                    <li class="menu-item">
+                        <a href="https://extrachill.com/category/interviews/">Interviews</a>
                     </li>
                     <li class="menu-item">
-                        <a href="https://community.extrachill.com/r/music-discussion">Music Discussion</a>
+                        <a href="https://extrachill.com/category/live-music-reviews/">Live Reviews</a>
                     </li>
                     <li class="menu-item">
-                        <a href="https://community.extrachill.com/r/music-discussion/music-festivals">Music Festivals</a>
+                        <a href="https://extrachill.com/festival-wire">Festival Wire</a>
                     </li>
                     <li class="menu-item">
-                        <a href="https://community.extrachill.com/r/music-discussion/music-industry">Music Industry</a>
+                        <a href="https://extrachill.com/category/song-meanings/">Song Meanings</a>
+                    </li>
+                    <li class="menu-item">
+                        <a href="https://extrachill.com/category/music-news/">Music News</a>
                     </li>
                 </ul>
             </li>
@@ -96,15 +68,6 @@
                 <ul class="sub-menu">
                     <li class="menu-item">
                         <a href="https://extrachill.com/contact/">Contact Us</a>
-                    </li>
-                    <li class="menu-item">
-                        <a href="https://extrachill.com/about/the-history-of-extra-chill/">History of Extra Chill</a>
-                    </li>
-                    <li class="menu-item">
-                        <a href="https://extrachill.com/festival/">Extra Chill Fest</a>
-                    </li>
-                    <li class="menu-item">
-                        <a href="https://extrachill.com/category/extra-chill-presents/">Extra Chill Presents</a>
                     </li>
                     <li class="menu-item">
                         <a href="https://extrachill.com/about/in-the-press/">In the Press</a>